multi: refactor so OfferHandler can access MessengerUtilities
